Refactored CommentBox into dumb and logic components.
