+++ src/sidebar-overlay/CommentBox/index.js
@@ -1,7 +1,7 @@
 import * as actions from './actions'
 import * as selectors from './selectors'
 import reducer from './reducer'
-import CommentBox from './components/CommentBox'
+import CommentBox from './components/container'
 
 export { actions, selectors, reducer }
 export default CommentBox
